+++ app/sprinkles/account/src/Database/Migrations/v440/PrimaryIdpsTable.php
@@ -1,0 +1,59 @@
+<?php
+
+/*
+ * UserFrosting (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/UserFrosting
+ * @copyright Copyright (c) 2019 Alexander Weissman
+ * @license   https://github.com/userfrosting/UserFrosting/blob/master/LICENSE.md (MIT License)
+ */
+
+namespace UserFrosting\Sprinkle\Account\Database\Migrations\v440;
+
+use Illuminate\Database\Schema\Blueprint;
+use UserFrosting\Sprinkle\Core\Database\Migration;
+
+/**
+ * Primary Identity Providers migration.
+ * Manages primary identity providers authentication information.
+ * Version 4.4.0.
+ *
+ * See https://laravel.com/docs/5.8/migrations#tables
+ *
+ * @author Archey Barrell
+ * @author Amos Folz
+ */
+class PrimaryIdpsTable extends Migration
+{
+    /**
+     * {@inheritdoc}
+     */
+    public static $dependencies = [
+    ];
+
+    /**
+     * {@inheritdoc}
+     */
+    public function up()
+    {
+        if (!$this->schema->hasTable('primary_idps')) {
+            $this->schema->create('primary_idps', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('slug')->unique();
+                $table->timestamps();
+
+                $table->engine = 'InnoDB';
+                $table->collation = 'utf8_unicode_ci';
+                $table->charset = 'utf8';
+            });
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function down()
+    {
+        $this->schema->drop('primary_idps');
+    }
+}